Move `test_graph_ql_error_type` to mod `error`
